cleanup; move functions into pixgroup.py
